// frontend/src/app/share/model/todo-Info.ts
export class TodoInfo {
  username:string;
  nickname:string;
  creationTime:string;
  avatarUrl:string;
  email:string;
  id: number;
  userId: number;
  title: string;
  content: string;
  beginDate: string;
  endDate: string;
  remindDate: string;
  categoryId: string;
  isCompleted: number;
  isImportant: number;
  isFavorite: number;
}